Chakra UI Setup

+++ backend/controllers/product.controller.js
@@ -63,6 +63,13 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product ID" }); // Validate the ID format
+  }
+  
   try {
     const deletedProduct = await Product.findByIdAndDelete(id);
     if (!deletedProduct) {
